[Core] Move MSBuild process runner logic out of NuGet addin into Core
Added an MSBuildProcessService which can be used to directly run
msbuild.
